More changes for moving to root
